backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
alembic
redis
python-multipart
python-jose[cryptography]
passlib[argon2]
pydantic-settings
minio
websockets
httpx
emails
aiosmtplib
jinja2
google-auth
google-auth-oauthlib
google-auth-httplib2
authlib
python-dotenv
